requirements.txt
pydantic>=2.7.4,<3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
sentence-transformers==2.6.0
chromadb==0.6.3
python-dotenv==1.0.0
pyyaml==6.0.1
nltk==3.8.1
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
redis==5.0.1
pypdf==3.17.1
docx2txt==0.8
langchain_chroma
langchain_huggingface
rank_bm25
